Begin Channels backend.
... and in the process, realize that the `pop_task()` method on the
backend is not applicable to the Channels backend. Remove it for now and
consider adding it back when/if we add other backends.

It's not applicable because we'll use the Django Channels router to
listen and process incoming tasks. We could write our own (simpler)
version of this by reaching into ChannelLayer's internals.