Add dummy-task-runner component.
A simple Vuejs component that starts a task, then polls the server to
find out the current status of the task. When the task has finished
(canceled/failed/completed), then the user can start a new task to
replace it. Progress is currently shown by dumping the JSON response we
get from the server.

This code uses ES6/7 features which aren't supported by all browsers.
Only tested in Chrome 55, and I expect it'll only work in Chrome 55 or
Edge, due to the use of `await`/`async`. (If this were a serious
project, I'd bother setting up Babel, honest!)